% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{BlWh.colors}
\alias{BlWh.colors}
\title{Color palette \code{BlWh}}
\usage{
BlWh.colors(n = 256)
}
\arguments{
\item{n}{Number of gradations between blue and white.}
}
\value{
Sequence of colors with blue and white at the ends.
The length is \code{n}.
}
\description{
This function generates a color palette consisting of blue (low) and white (high).
}
