% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{cov.multiple.edge}
\alias{cov.multiple.edge}
\title{Covariance of multiple edge test statistics}
\usage{
cov.multiple.edge(W.S, S.T, ns, E.set = NULL, E.mat = NULL)
}
\arguments{
\item{W.S}{List with length \code{m} of 2-dimensional array (\code{q*q}) or 3-dimensional array (\code{q*q*num.lambda}) containing the spatial precision or partial-correlation matrices over \code{m} sessions.}

\item{S.T}{List with length \code{m} of 2-dimensional array (\code{p*p}) containing the temporal covariance matrices over \code{m} sessions.}

\item{ns}{Number of trials. A numeric for all sessions or a sequence of a different order for each session can be given.}

\item{E.set}{2-dimensional integer array of size \code{num.E*2} containing the pairs of index for the nodes the target edges connect. If not provided, it is postulated from argument \code{E.mat}.}

\item{E.mat}{2-dimensional boolen array of size \code{q*q} containing the adjacency matrix of the target edges. if not provided, it is postulated from arguemnt \code{E.set}.}
}
\value{
Estimated covariance matrix for the test statistics
}
\description{
This function calculates the covariance matrix of test statistics for multiple edge
inference.
}
