% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{est.spatial}
\alias{est.spatial}
\title{Estimate spatial covariance matrices}
\usage{
est.spatial(data, lambdas = NULL, eps = 1e-04, maxit = 10000, verbose = FALSE)
}
\arguments{
\item{data}{List with length \code{m} of 3-dimensional array (\code{p*q*n.l}) containing \code{n.l} matrix observations of dimension \code{p*q} over sessions \code{l} = 1, ..., m.}

\item{lambdas}{Sequence of group lasso penalty hyperparameters. Default is determined by \code{p}, \code{q}, \code{m}, and \code{n.l}'s.}

\item{eps}{Threshold value for convergence. Default is 1e-4.}

\item{maxit}{Maximum number of iteration. Default is 10,000.}

\item{verbose}{Boolean with default \code{FALSE}.}
}
\value{
List output containing the spatial estimation result

\code{W.hat.S} List of Omega.hat.S, the estimate of the spatial precision matrix, at each session.

\code{S.hat.S} List of Sigma.hat.S, the estimate of the spatial covariance matrix, at each session.

\code{b.hat.S} List of beta.hat.S, the estimate of the spatial regression parameters, at each session.

\code{P.hat.S} List of Phi.hat.S, the estimate of the spatial residual covariance, at eachs ession.
}
\description{
This function estimates the spatial covariance parameter from
matrix-variate observation data over multiple sessions.
}
