% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{est.temporal}
\alias{est.temporal}
\title{Estimate temporal covariance matrices}
\usage{
est.temporal(data, order.T = NULL, decay.T = NULL, verbose = FALSE)
}
\arguments{
\item{data}{List with length \code{m} of 3-dimensional array (\code{p*q*n}) containing \code{n} matrix observations of dimension \code{p*q} over \code{m} sessions}

\item{order.T}{Order for the temporal AR models. A numeric for all sessions or a sequence of a different order for each session can be given. If not provided, the default value is determined using `decay.T}

\item{decay.T}{Decay rate of AR parameter. A numeric for all sessions or a sequence of a different rate for each session can be given. Default is 0.5.}

\item{verbose}{Boolean with default \code{FALSE}}
}
\value{
List output containing the temporal estimation result

\code{W.hat.T} List of Omega.hat.T, the estimate of the temporal precision matrix, at each session.

\code{S.hat.T} List of Sigma.hat.T, the estimate of the temporal covariance matrix, at each session.

\code{b.hat.T} List of beta.hat.T, the estimate of the temporal regression parameters, at each session.

\code{P.hat.T} List of Phi.hat.T, the estimate of the temporal residual variances, at each session
}
\description{
This function estimates the temporal covariance parameter from
matrix-variate observation data over multiple sessions.
}
