% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{imshow}
\alias{imshow}
\title{Image plot of matrix data}
\usage{
imshow(x, vmin = NULL, vmax = NULL, col = RdBl.colors(256))
}
\arguments{
\item{x}{Matrix data input}

\item{vmin}{Minimum value of the colorscale. If not provided, the default is \code{-vmax}.}

\item{vmax}{Maximum value of the colorscale. If not provided, the default is the maximum absolute value of \code{x}.}

\item{col}{Color palette at which the image is plotted.}
}
\value{
Imageplot of the input in the matrix position.
}
\description{
Wrapper function of \code{heatmap} having a similar interface with \code{plt.imshow} in Python.
}
