% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inf.single.edge}
\alias{inf.single.edge}
\title{Single edge inference}
\usage{
inf.single.edge(W.hat.S, S.hat.T, ns)
}
\arguments{
\item{W.hat.S}{List with length \code{m} of 2-dimensional array (\code{q*q}) or 3-dimensional array (\code{q*q*num.lambda}) containing the spatial precision matrix estimates over \code{m} sessions.}

\item{S.hat.T}{List with length \code{m} of 2-dimensional array (\code{p*p}) containing the temporal covariance matrix estimates over \code{m} sessions.}

\item{ns}{Number of trials. A numeric for all sessions or a sequence of a different order for each session can be given.}
}
\value{
List output containing the inference result

\code{R.hat.S} List of the estimated partial correlation matrix at each session.

\code{p.value} The p-values for the entries of the precision matrix to be zero at each session.

\code{T.hat} Estimated test statistics \

\code{Var.T.hat} Estimated variances for the test statistics
}
\description{
This function performs statistical tests on edge-wise null hypotheses.
}
