% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{RdBl.colors}
\alias{RdBl.colors}
\title{Color palette \code{RdBl}}
\usage{
RdBl.colors(n = 128)
}
\arguments{
\item{n}{Number of gradations between red and white and between white and blue.}
}
\value{
Sequence of colors with red and blue at the ends and white at the middle.
The length is \code{2*n+1}.
}
\description{
This function generates a color palette consisting of red (low), white (mid), and
blue (high).
}
