% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{WhBl.colors}
\alias{WhBl.colors}
\title{Color palette \code{WhBl}}
\usage{
WhBl.colors(n = 256)
}
\arguments{
\item{n}{Number of gradations between white and blue.}
}
\value{
Sequence of colors with blue and white at the ends.
The length is \code{n}.
}
\description{
This function generates a color palette consisting of white (low) and blue (high).
}
