% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inf.multiple.edge}
\alias{inf.multiple.edge}
\title{Multiple edge inference}
\usage{
inf.multiple.edge(
  W.hat.S,
  S.hat.T,
  ns,
  E.set = NULL,
  E.mat = NULL,
  num.bst = 10000
)
}
\arguments{
\item{W.hat.S}{List with length \code{m} of 2-dimensional array (\code{q*q}) or 3-dimensional array (\code{q*q*num.lambda}) containing the spatial precision matrix estimates over \code{m} sessions.}

\item{S.hat.T}{List with length \code{m} of 2-dimensional array (\code{p*p}) containing the temporal covariance matrix estimates over \code{m} sessions.}

\item{ns}{Number of trials. A numeric for all sessions or a sequence of a different order for each session can be given.}
}
\value{
List of output containing the inference result.

\code{R.hat.S} List of the estimated partial correlation matrix at each session.

\code{p.value} p-values for the global null hypothesis

\code{T.hat.E} Estimated test statistics

\code{Cov.T.E} Estimated covariance matrix for the test statistics
}
\description{
This function performs a statistical test on a global null hypothesis over a given
edge set.
}
